School completed

+++ src/modules/brand/brand.service.js
@@ -1,6 +1,10 @@
 import { ResData } from "../../common/resData.js";
 import { BrandRepository } from "./brand.repository.js";
 import { BrandEntity } from "./entity/brand.entity.js";
+import {
+  BrandAlreadyExistFoundByName,
+  BrandNotFoundById,
+} from "./exception/brand.exception.js";
 
 export class BrandService {
   #repository;
@@ -8,6 +12,12 @@ export class BrandService {
     this.#repository = new BrandRepository();
   }
   async create(dto) {
+    const foundBrand = await this.#repository.findByName(dto.name);
+
+    if (foundBrand) {
+      throw new BrandAlreadyExistFoundByName();
+    }
+
     const newBrand = new BrandEntity(dto);
 
     const brand = await this.#repository.insert(newBrand);
@@ -27,12 +37,22 @@ export class BrandService {
   async getById(brandId) {
     const foundBrand = await this.#repository.findOneById(brandId);
 
+    if (!foundBrand) {
+      throw new BrandNotFoundById();
+    }
+
     const resData = new ResData("Found brand by id", 200, foundBrand);
 
     return resData;
   }
 
   async delete(brandId) {
+    const foundBrand = await this.#repository.findOneById(brandId);
+
+    if (!foundBrand) {
+      throw new BrandNotFoundById();
+    }
+
     const deletedBrand = await this.#repository.delete(brandId);
 
     const resData = new ResData("Deleted brand by id", 200, deletedBrand);
